Moved SubproblemStatus.hpp to uno/ingredients/subproblem_solvers/
