+++ CMakeLists.txt
@@ -53,7 +53,6 @@ file(GLOB UNO_SOURCE_FILES
    uno/ingredients/inequality_handling_methods/inequality_constrained_methods/*.cpp
    uno/ingredients/inequality_handling_methods/interior_point_methods/*.cpp
    uno/ingredients/subproblem_solvers/*.cpp
-   uno/ingredients/subproblems/*.cpp
    uno/model/*.cpp
    uno/optimization/*.cpp
    uno/options/*.cpp
